+++ packages/shared/src/ShadowRoot/createReactRootShadowed.tsx
